refactor service config generator to work with new base class
